delete utils.py, move function to common.py
